agent/requirements.txt
# keep in alphabetical order
dm_control
omegaconf==2.3.0
Pillow
pin==3.7.0
pygame
pyspacemouse
PyQt6
pyquaternion
pyrealsense2
pyzmq
pure-python-adb
quaternion
setuptools
tyro
ur-rtde
zmq
xarm
xarm-python-sdk
numpy-quaternion
termcolor
flask_cors==5.0.0
